improved: by adopting MVVM with Combine to manage Photo from Unsplash API (#10)
