+++ UIKit-ImageCollectionView/ViewControllers/TabBarViewController.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SnapKit
 
 class TabBarViewController: UITabBarController {
